+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 BUG FIXES:
 * vpc: Security group rule port bugfix: can create rules without specifying a port
 * vpc: Fix internal_address_spec block in network load balancer resource doc
+* dataproc: support for UI Proxy
 
 ## 0.45.1 (October 06, 2020)